Feature/py to raster (#49)
* Add histogram methods and tests
* backup commit before rewriting discrete tile
* add fracTile and test to itemCount()
* add requirements change test
* Modify tests
* Add tests, improve legibility, cite code reuse
* Add toIntRaster and toDoubleRaster methods
Add these methods to pyrasterframes, piggybacking off of the existing toRaster method in RasterFrameMethods.scala. Change types.py and pyRFContext to give access to these methods, which will probbaly be used primarily for plotting images in python.
* add noDataFilter
Implementation could still be incorrect as the test has failed. This is an intermediate commit.
* reverted wrong branch commit
* fix docker-compose
* remove noDataFilter, add pydocs

Signed-off-by: James Nachbar <jnachbar@astraea.io>